Tweak script

+++ scripts/redbox-rifcs.sh
@@ -39,7 +39,11 @@ fi
 
 mkdir -p rifcs/processed
 pushd rifcs
-wget -O jcu-rifcs.rif  https://eresearch.jcu.edu.au/tdh/data/@@rifcs?days_in_past=all
-$RIF_PYTHON ../TDH-Research-Data-Catalogue/src/main/python/bin/rifsplit.py jcu-rifcs.rif
-$RIF_PYTHON ../TDH-Research-Data-Catalogue/src/main/python/bin/process_rif_cs.py -o processed output/collection*.rif
-
+pwd
+if [ ! -e jcu-rifcs.rif ]
+then
+    wget -O jcu-rifcs.rif  https://eresearch.jcu.edu.au/tdh/data/@@rifcs?days_in_past=all
+fi
+$RIF_PYTHON ../TDH-Research-Data-Catalogue.git/src/main/python/bin/rifsplit.py jcu-rifcs.rif
+$RIF_PYTHON ../TDH-Research-Data-Catalogue.git/src/main/python/bin/process_rif_cs.py -o processed output/collection*.rif
+popd
